Добавление моделей.

+++ medservices/models/service.py
@@ -2,4 +2,4 @@ from django.db import models
 
 class Service(models.Model):
     serv_name = models.TextField('serv_name')
-    serv_cost = models.IntegerField('serv_cost')
+    serv_cost = models.DecimalField('serv_cost', max_digits=7, decimal_places=2)
